TEE Benchmark PoC
Add fixes after review
Add support for allocate ringbuf array outside TEEC_InvokeCommand
for performing benchmarks
